// ui/src/widgets/Layout/model/types.ts
import type { ElementType } from 'react'

import type { LinkProps } from 'react-router-dom'

type User = {
  name: string
  email: string
  avatar: string
}

type Team = {
  name: string
  logo: ElementType
  plan: string
}

type BaseNavItem = {
  title: string
  badge?: string
  icon?: ElementType
}

type NavLink = BaseNavItem & {
  url: LinkProps['to'] | (string & {})
  items?: never
}

type NavCollapsible = BaseNavItem & {
  items: (BaseNavItem & { url: LinkProps['to'] | (string & {}) })[]
  url?: never
}

type NavItem = NavCollapsible | NavLink

type NavGroup = {
  title: string
  items: NavItem[]
}

type SidebarData = {
  user: User
  teams: Team[]
  navGroups: NavGroup[]
}

export type { SidebarData, NavGroup, NavItem, NavCollapsible, NavLink }